Move DnsMessage to okhttp core internal
