+++ okhttp-dnsoverhttps/src/test/java/okhttp3/dnsoverhttps/DnsOverHttpsTest.kt
@@ -52,14 +52,14 @@ import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import okhttp3.Protocol
 import okhttp3.Response
-import okhttp3.dnsoverhttps.internal.CLASS_IN
-import okhttp3.dnsoverhttps.internal.DnsMessage
-import okhttp3.dnsoverhttps.internal.Question
-import okhttp3.dnsoverhttps.internal.ResourceRecord
-import okhttp3.dnsoverhttps.internal.TYPE_A
-import okhttp3.dnsoverhttps.internal.TYPE_AAAA
+import okhttp3.internal.dns.CLASS_IN
 import okhttp3.internal.dns.DnsEvent
+import okhttp3.internal.dns.DnsMessage
 import okhttp3.internal.dns.EntryPoint
+import okhttp3.internal.dns.Question
+import okhttp3.internal.dns.ResourceRecord
+import okhttp3.internal.dns.TYPE_A
+import okhttp3.internal.dns.TYPE_AAAA
 import okhttp3.internal.dns.invoke
 import okhttp3.internal.dns.toEventsQueue
 import okhttp3.testing.PlatformRule
