+++ okhttp-dnsoverhttps/src/test/java/okhttp3/dnsoverhttps/DnsRecordCodecTest.kt
@@ -22,13 +22,13 @@ import assertk.assertions.isEqualTo
 import java.net.InetAddress
 import java.net.UnknownHostException
 import kotlin.test.assertFailsWith
-import okhttp3.dnsoverhttps.internal.DnsMessage
-import okhttp3.dnsoverhttps.internal.DnsMessageReader
-import okhttp3.dnsoverhttps.internal.RESPONSE_CODE_SUCCESS
-import okhttp3.dnsoverhttps.internal.ResourceRecord
-import okhttp3.dnsoverhttps.internal.TYPE_A
-import okhttp3.dnsoverhttps.internal.TYPE_AAAA
 import okhttp3.dnsoverhttps.internal.asQueryParameter
+import okhttp3.internal.dns.DnsMessage
+import okhttp3.internal.dns.DnsMessageReader
+import okhttp3.internal.dns.RESPONSE_CODE_SUCCESS
+import okhttp3.internal.dns.ResourceRecord
+import okhttp3.internal.dns.TYPE_A
+import okhttp3.internal.dns.TYPE_AAAA
 import okio.Buffer
 import okio.ByteString
 import okio.ByteString.Companion.decodeHex
