+++ okhttp-dnsoverhttps/src/test/java/okhttp3/dnsoverhttps/internal/DnsMessageReaderRecordedValuesTest.kt
@@ -20,6 +20,12 @@ import assertk.assertions.containsExactly
 import assertk.assertions.isEqualTo
 import java.net.InetAddress
 import kotlin.test.Test
+import okhttp3.internal.dns.DnsMessage
+import okhttp3.internal.dns.DnsMessageReader
+import okhttp3.internal.dns.Question
+import okhttp3.internal.dns.ResourceRecord
+import okhttp3.internal.dns.TYPE_A
+import okhttp3.internal.dns.TYPE_AAAA
 import okio.Buffer
 import okio.ByteString.Companion.decodeHex
