+++ okhttp-testing-support/src/main/kotlin/okhttp3/FakeDns.kt
@@ -28,16 +28,16 @@ import java.util.concurrent.atomic.AtomicInteger
 import mockwebserver3.Dispatcher
 import mockwebserver3.MockResponse
 import mockwebserver3.RecordedRequest
-import okhttp3.dnsoverhttps.internal.CLASS_IN
-import okhttp3.dnsoverhttps.internal.DnsMessage
-import okhttp3.dnsoverhttps.internal.DnsMessageReader
 import okhttp3.dnsoverhttps.internal.DnsMessageWriter
-import okhttp3.dnsoverhttps.internal.Question
-import okhttp3.dnsoverhttps.internal.ResourceRecord
-import okhttp3.dnsoverhttps.internal.TYPE_A
-import okhttp3.dnsoverhttps.internal.TYPE_AAAA
-import okhttp3.dnsoverhttps.internal.TYPE_HTTPS
 import okhttp3.internal.concurrent.TaskRunner
+import okhttp3.internal.dns.CLASS_IN
+import okhttp3.internal.dns.DnsMessage
+import okhttp3.internal.dns.DnsMessageReader
+import okhttp3.internal.dns.Question
+import okhttp3.internal.dns.ResourceRecord
+import okhttp3.internal.dns.TYPE_A
+import okhttp3.internal.dns.TYPE_AAAA
+import okhttp3.internal.dns.TYPE_HTTPS
 import okio.Buffer
 import okio.ByteString.Companion.decodeBase64
