+++ okhttp/src/commonJvmAndroid/kotlin/okhttp3/internal/dns/-DnsMessage.kt
@@ -15,20 +15,10 @@
  */
 @file:Suppress("ktlint:standard:filename")
 
-package okhttp3.dnsoverhttps.internal
+package okhttp3.internal.dns
 
-import java.io.IOException
 import java.net.InetAddress
-import java.net.ProtocolException
-import okhttp3.Protocol
-import okhttp3.RequestBody
-import okhttp3.Response
-import okhttp3.dnsoverhttps.DnsOverHttps.Companion.DNS_MESSAGE
-import okhttp3.dnsoverhttps.DnsOverHttps.Companion.MAX_RESPONSE_SIZE
 import okhttp3.internal.OkHttpInternalApi
-import okhttp3.internal.platform.Platform
-import okio.Buffer
-import okio.BufferedSink
 import okio.ByteString
 
 @OkHttpInternalApi
@@ -166,46 +156,3 @@ internal const val SERVICE_PARAMETER_PORT = 3
 internal const val SERVICE_PARAMETER_IPV4_HINT = 4
 internal const val SERVICE_PARAMETER_ECH = 5
 internal const val SERVICE_PARAMETER_IPV6_HINT = 6
-
-internal fun DnsMessage.asQueryParameter(): String {
-  val buffer = Buffer()
-  DnsMessageWriter(buffer).write(this@asQueryParameter)
-  return buffer.readByteString().base64Url().replace("=", "")
-}
-
-internal class QueryRequestBody(
-  private val query: DnsMessage,
-) : RequestBody() {
-  override fun contentType() = DNS_MESSAGE
-
-  override fun writeTo(sink: BufferedSink) {
-    DnsMessageWriter(sink.buffer).write(query)
-    sink.emitCompleteSegments()
-  }
-}
-
-@Throws(IOException::class)
-internal fun decodeResponse(response: Response): DnsMessage {
-  if (
-    response.cacheResponse == null &&
-    response.protocol !== Protocol.HTTP_2 &&
-    response.protocol !== Protocol.QUIC
-  ) {
-    Platform.get().log("Unexpected protocol: ${response.protocol}", Platform.WARN)
-  }
-
-  response.use {
-    if (!response.isSuccessful) {
-      throw IOException("response: ${response.code} ${response.message}")
-    }
-
-    val body = response.body
-    if (body.contentLength() > MAX_RESPONSE_SIZE) {
-      throw ProtocolException(
-        "response size exceeds limit ($MAX_RESPONSE_SIZE bytes): ${body.contentLength()} bytes",
-      )
-    }
-
-    return DnsMessageReader(body.source()).read()
-  }
-}
